+++ src/auth/__init__.py
@@ -1,2 +1,2 @@
-from .auth import authenticate, identify
+from .auth import AuthHandler
 from .validators import validate_sign_up
